Add GitHub action to run papermill and push updated data.
This commit adds a GitHub action which runs the data update pipeline
when the feature/automated-data-update branch is pushed to as well as
on a 30 minute schedule (once this commit is in master).

The action will run essentially the same proccess as
'docker/process-data', which runs papermill against notebooks who's
data might update. These notebooks will publish to the data/processed
and  data/published folders. If this produces new data, those changes
- along with the executed notebooks that serve as logs - will be
committed and pushed to the 'data/update' branch.